# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenny Live API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PanelRequest {
  /** The product/career question to debate */
  question: string;
  /**
   * Number of panelists (2 or 3)
   * @minimum 2
   * @maximum 3
   */
  panelSize?: number;
}

export interface ManualPanelRequest {
  /** The product/career question to debate */
  question: string;
  /**
   * Names of guests to include in the panel
   * @minItems 2
   * @maxItems 3
   */
  guestNames: string[];
}

export interface GuestInfo {
  name: string;
  stance: string;
  episode: string;
  title: string;
  timestamp: string;
  color: string;
  relevantChunks?: string[];
}

export interface PanelResponse {
  guests: GuestInfo[];
}

export interface GuestSummary {
  name: string;
  filename: string;
  title: string;
}

export interface GuestListResponse {
  guests: GuestSummary[];
}