# lib/api-zod/src/generated/types/guestInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenny Live API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GuestInfo {
  name: string;
  stance: string;
  episode: string;
  title: string;
  timestamp: string;
  color: string;
  relevantChunks?: string[];
}