# lib/api-zod/src/generated/types/guestListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenny Live API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuestSummary } from "./guestSummary";

export interface GuestListResponse {
  guests: GuestSummary[];
}