# lib/api-zod/src/generated/types/guestSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenny Live API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GuestSummary {
  name: string;
  filename: string;
  title: string;
}