# lib/api-zod/src/generated/types/manualPanelRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenny Live API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ManualPanelRequest {
  /** The product/career question to debate */
  question: string;
  /**
   * Names of guests to include in the panel
   * @minItems 2
   * @maxItems 3
   */
  guestNames: string[];
}