# lib/api-zod/src/generated/types/panelRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenny Live API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PanelRequest {
  /** The product/career question to debate */
  question: string;
  /**
   * Number of panelists (2 or 3)
   * @minimum 2
   * @maximum 3
   */
  panelSize?: number;
}